Refactor: Flatten database module
Flattened relvar-core/src/database/* into a single relvar-core/src/database.rs file to enforce KISS and remove unnecessary layer lasagna.
